Move bulk tests to seperate file.
